@@ -11,52 +11,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.0.0",
-    "@angular/common": "~7.0.0",
-    "@angular/compiler": "~7.0.0",
-    "@angular/core": "~7.0.0",
-    "@angular/forms": "~7.0.0",
-    "@angular/http": "~7.0.0",
-    "@angular/platform-browser": "~7.0.0",
-    "@angular/platform-browser-dynamic": "~7.0.0",
-    "@angular/router": "~7.0.0",
+    "@angular/animations": "~8.2.3",
+    "@angular/common": "~8.2.3",
+    "@angular/compiler": "~8.2.3",
+    "@angular/core": "~8.2.3",
+    "@angular/forms": "~8.2.3",
+    "@angular/http": "^7.2.15",
+    "@angular/platform-browser": "~8.2.3",
+    "@angular/platform-browser-dynamic": "~8.2.3",
+    "@angular/router": "~8.2.3",
     "@sweetalert2/ngx-sweetalert2": "^5.1.0",
-    "@swimlane/ngx-datatable": "^14.0.0",
-    "angular-admin-lte": "2.0.1",
+    "angular-admin-lte": "^2.0.1",
     "angular-loading-page": "^0.6.0",
-    "bootstrap": "^4.1.3",
-    "bootstrap-css-only": "3.3.6",
-    "ckeditor4-angular": "1.0.0-beta.2",
-    "core-js": "^2.5.4",
-    "font-awesome": "^4.7.0",
-    "ionicons": "^4.4.7",
-    "ngx-bootstrap": "^3.1.2",
-    "ngx-toastr": "^9.1.1",
-    "node-sass": "^4.11.0",
-    "prismjs": "^1.15.0",
-    "rxjs": "~6.3.3",
-    "sweetalert2": "^8.15.1",
-    "zone.js": "~0.8.26"
+    "bootstrap-css-only": "3.3.7",
+    "ckeditor4-angular": "^1.0.0-beta.2",
+    "font-awesome": "4.7.0",
+    "ionicons": "2.0.1",
+    "ngx-bootstrap": "^5.1.1",
+    "ngx-toastr": "^10.1.0",
+    "prismjs": "^1.17.1",
+    "rxjs": "~6.4.0",
+    "sweetalert2": "^8.16.3",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.10.6",
-    "@angular/cli": "~7.0.3",
-    "@angular/compiler-cli": "~7.0.0",
-    "@angular/language-service": "~7.0.0",
-    "@types/jasmine": "~2.8.8",
+    "@angular-devkit/build-angular": "~0.803.1",
+    "@angular/cli": "~8.3.1",
+    "@angular/compiler-cli": "~8.2.3",
+    "@angular/language-service": "~8.2.3",
+    "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.1"
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
   }
 }
