@@ -10,14 +10,14 @@ export class ClientModuleService {
 
   // 创建客户端模块
   public CreateAsync(vm: Rainbow.ViewModels.ClientModules.CreateClientModuleVM)
-    : Observable<Yunyong.Core.AsyncTaskTResult<String>> {
-    return this.http.post<Yunyong.Core.AsyncTaskTResult<String>>
+    : Observable<Yunyong.Core.AsyncTaskTResult<string>> {
+    return this.http.post<Yunyong.Core.AsyncTaskTResult<string>>
       (`${this.baseUrl}api/ClientModule/Create`, vm, getHttpOptions());
   }
   // 更新客户端模块
   public UpdateAsync(vm: Rainbow.ViewModels.ClientModules.UpdateClientModuleVM)
-    : Observable<Yunyong.Core.AsyncTaskTResult<String>> {
-    return this.http.put<Yunyong.Core.AsyncTaskTResult<String>>
+    : Observable<Yunyong.Core.AsyncTaskTResult<string>> {
+    return this.http.put<Yunyong.Core.AsyncTaskTResult<string>>
       (`${this.baseUrl}api/ClientModule/Update`, vm, getHttpOptions());
   }
   // 查询客户端模块列表（分页）
@@ -27,7 +27,7 @@ export class ClientModuleService {
       (`${this.baseUrl}api/ClientModule/Query?${stringify(option)}`, getHttpOptions());
   }
   // 获取客户端模块
-  public GetAsync(id: String)
+  public GetAsync(id: string)
     : Observable<Rainbow.ViewModels.ClientModules.ClientModuleVM> {
     return this.http.get<Rainbow.ViewModels.ClientModules.ClientModuleVM>
       (`${this.baseUrl}api/ClientModule?${stringify(id)}`, getHttpOptions());
