@@ -15,13 +15,13 @@ export class MessageService {
       (`${this.baseUrl}api/Message/Query?${stringify(option)}`, getHttpOptions());
   }
   // GetAsync
-  public GetAsync(msgId: String)
+  public GetAsync(msgId: string)
     : Observable<Rainbow.ViewModels.Messages.MessageVM> {
     return this.http.get<Rainbow.ViewModels.Messages.MessageVM>
       (`${this.baseUrl}api/Message/Get/${msgId}`, getHttpOptions());
   }
   // ReadedAsync
-  public ReadedAsync(msgId: String)
+  public ReadedAsync(msgId: string)
     : Observable<Yunyong.Core.AsyncTaskResult> {
     return this.http.put<Yunyong.Core.AsyncTaskResult>
       (`${this.baseUrl}api/Message/Readed/${msgId}`, msgId, getHttpOptions());
